# ms/controllerblueprints/application/src/main/resources/swagger-ui/dist/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.3.0-SNAPSHOT",
    "title" : "Controller Blueprint Service"
  },
  "basePath" : "/api/controller-blueprints/v1",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/service/dictionary" : {
      "post" : {
        "summary" : "Provides Rest service to Save Resource dictionary Type",
        "description" : "",
        "operationId" : "saveResourceDictionary",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ResourceDictionary"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResourceDictionary"
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/dictionary/{name}" : {
      "get" : {
        "summary" : "Provides Rest service to get Resource dictionary",
        "description" : "",
        "operationId" : "getResourceDictionaryByName",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResourceDictionary"
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      },
      "delete" : {
        "summary" : "Provides Rest service to delete ResourceDictionary Type",
        "description" : "",
        "operationId" : "deleteResourceDictionaryByName",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/dictionarybynames" : {
      "post" : {
        "summary" : "Provides Rest service to get ResourceDictionary Type by names",
        "description" : "",
        "operationId" : "searchResourceDictionaryByNames",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResourceDictionary"
              }
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/dictionarysearch/{tags}" : {
      "get" : {
        "summary" : "Provides Rest service to search Resource dictionary by tags",
        "description" : "",
        "operationId" : "searchResourceDictionaryByTags",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "tags",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResourceDictionary"
              }
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
...
    "/service/resourceassignment/automap" : {
      "post" : {
        "summary" : "Provides Rest service to auto map for the Resource assignments",
        "description" : "",
        "operationId" : "autoMap",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ResourceAssignment"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AutoMapResponse"
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/resourceassignment/generate" : {
      "post" : {
        "summary" : "Provides Rest service to auto map for the Resource Mapping",
        "description" : "",
        "operationId" : "generateResourceAssignments",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ConfigModelContent"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResourceAssignment"
              }
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/resourceassignment/validate" : {
      "post" : {
        "summary" : "Provides Rest service to validate Resource assignments",
        "description" : "",
        "operationId" : "validateResourceAssignments",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ResourceAssignment"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResourceAssignment"
              }
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/servicetemplate/enrich" : {
      "post" : {
        "summary" : "Provides Rest service to enrich service template",
        "description" : "",
        "operationId" : "enrichServiceTemplate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ServiceTemplate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ServiceTemplate"
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    },
    "/service/servicetemplate/validate" : {
      "post" : {
        "summary" : "Provides Rest service to validate service template",
        "description" : "",
        "operationId" : "validateServiceTemplate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ServiceTemplate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ServiceTemplate"
            }
          },
          "404" : {
            "description" : "Service not available"
          },
          "500" : {
            "description" : "Unexpected Runtime error"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "ArtifactType" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "attributes" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/AttributeDefinition"
          }
        },
        "properties" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PropertyDefinition"
          }
        },
        "derived_from" : {
          "type" : "string"
        },
        "mime_type" : {
          "type" : "string"
        },
        "file_ext" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "AutoMapResponse" : {
      "type" : "object",
      "properties" : {
        "resourceAssignments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ResourceAssignment"
          }
        },
        "dataDictionaries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ResourceDictionary"
          }
        }
      }
    },
...
    "ConfigModel" : {
      "type" : "object",
      "required" : [ "artifactName", "artifactVersion", "published", "tags", "updatedBy" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "serviceUUID" : {
          "type" : "string"
        },
        "distributionId" : {
          "type" : "string"
        },
        "serviceName" : {
          "type" : "string"
        },
        "serviceDescription" : {
          "type" : "string"
        },
        "resourceUUID" : {
          "type" : "string"
        },
        "resourceInstanceName" : {
          "type" : "string"
        },
        "resourceName" : {
          "type" : "string"
        },
        "resourceVersion" : {
          "type" : "string"
        },
        "resourceType" : {
          "type" : "string"
        },
        "artifactUUId" : {
          "type" : "string"
        },
        "artifactType" : {
          "type" : "string"
        },
        "artifactVersion" : {
          "type" : "string"
        },
        "artifactDescription" : {
          "type" : "string"
        },
        "internalVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "artifactName" : {
          "type" : "string"
        },
        "published" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "tags" : {
          "type" : "string"
        },
        "configModelContents" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ConfigModelContent"
          }
        }
      }
    },
    "ConfigModelContent" : {
      "type" : "object",
      "required" : [ "content", "contentType", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "contentType" : {
          "type" : "string"
        },
        "configModel" : {
          "$ref" : "#/definitions/ConfigModel"
        },
        "description" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "ConstraintClause" : {
      "type" : "object",
      "properties" : {
        "equal" : {
          "type" : "object"
        },
        "greater_than" : {
          "type" : "object"
        },
        "greater_or_equal" : {
          "type" : "object"
        },
        "less_than" : {
          "type" : "object"
        },
        "less_or_equal" : {
          "type" : "object"
        },
        "in_range" : {
          "type" : "object"
        },
        "valid_values" : {
          "type" : "array",
          "items" : {
            "type" : "object"
          }
        },
        "length" : {
          "type" : "object"
        },
        "min_length" : {
          "type" : "object"
        },
        "max_length" : {
          "type" : "object"
        },
        "pattern" : {
          "type" : "string"
        }
      }
    },
...
    "DataType" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "attributes" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/AttributeDefinition"
          }
        },
        "properties" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PropertyDefinition"
          }
        },
        "constraints" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        },
        "derived_from" : {
          "type" : "string"
        }
      }
    },
...
    "ImportDefinition" : {
      "type" : "object",
      "properties" : {
        "file" : {
          "type" : "string"
        },
        "repository" : {
          "type" : "string"
        },
        "namespace_uri" : {
          "type" : "string"
        },
        "namespace_prefix" : {
          "type" : "string"
        }
      }
    },
...
    "NodeType" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "attributes" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/AttributeDefinition"
          }
        },
        "properties" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PropertyDefinition"
          }
        },
        "capabilities" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/CapabilityDefinition"
          }
        },
        "requirements" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/RequirementDefinition"
          }
        },
        "interfaces" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/InterfaceDefinition"
          }
        },
        "artifacts" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/ArtifactDefinition"
          }
        },
        "derived_from" : {
          "type" : "string"
        }
      }
    },
...
    "PolicyType" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "attributes" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/AttributeDefinition"
          }
        },
        "properties" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PropertyDefinition"
          }
        },
        "targets" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "derived_from" : {
          "type" : "string"
        }
      }
    },
...
    "PropertyDefinition" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "required" : {
          "type" : "boolean"
        },
        "type" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "constraints" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ConstraintClause"
          }
        },
        "value" : {
          "type" : "object"
        },
        "default" : {
          "type" : "object"
        },
        "entry_schema" : {
          "$ref" : "#/definitions/EntrySchema"
        }
      }
    },
...
    "RepositoryDefinition" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "credential" : {
          "$ref" : "#/definitions/Credential"
        }
      }
    },
...
    "ResourceAssignment" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "property" : {
          "$ref" : "#/definitions/PropertyDefinition"
        },
        "input-param" : {
          "type" : "boolean"
        },
        "dictionary-name" : {
          "type" : "string"
        },
        "dictionary-source" : {
          "type" : "string"
        },
        "dependencies" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "status" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        },
        "updated-date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updated-by" : {
          "type" : "string"
        }
      }
    },
    "ResourceDictionary" : {
      "type" : "object",
      "required" : [ "dataType", "definition", "description", "name", "resourcePath", "resourceType", "tags", "updatedBy" ],
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "resourcePath" : {
          "type" : "string"
        },
        "resourceType" : {
          "type" : "string"
        },
        "dataType" : {
          "type" : "string"
        },
        "entrySchema" : {
          "type" : "string"
        },
        "validValues" : {
          "type" : "string"
        },
        "sampleValue" : {
          "type" : "string"
        },
        "definition" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "tags" : {
          "type" : "string"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updatedBy" : {
          "type" : "string"
        }
      }
    },
    "ServiceTemplate" : {
      "type" : "object",
      "properties" : {
        "metadata" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "description" : {
          "type" : "string"
        },
        "repositories" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/RepositoryDefinition"
          }
        },
        "imports" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ImportDefinition"
          }
        },
        "tosca_definitions_version" : {
          "type" : "string"
        },
        "dsl_definitions" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "artifact_types" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/ArtifactType"
          }
        },
        "data_types" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/DataType"
          }
        },
        "node_types" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/NodeType"
          }
        },
        "policy_types" : {
          "$ref" : "#/definitions/PolicyType"
        },
        "topology_template" : {
          "$ref" : "#/definitions/TopologyTemplate"
        }
      }
    },
...
    "TopologyTemplate" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "inputs" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PropertyDefinition"
          }
        },
        "policies" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PolicyDefinition"
          }
        },
        "outputs" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/PropertyDefinition"
          }
        },
        "workflows" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/Workflow"
          }
        },
        "node_templates" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/NodeTemplate"
          }
        },
        "relationship_templates" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/RelationshipTemplate"
          }
        },
        "substitution_mappings" : {
          "type" : "object"
        }
      }
    },
...
  }
}